+++ transport/src/main/java/com/cooperative/transport/repositories/SalaireRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import java.util.List;
 @Repository
-public interface SalaireRepository extends JpaRepository<Salaires, Long> {
+public interface SalaireRepository extends JpaRepository<Salaires, Integer> {
     @Query("Select s from Salaires s where s.employe.id=:id order by s.date_modification desc limit 1")
-    Salaires findByIdEmp(@Param("id") Long id);
+    Salaires findByIdEmp(@Param("id") Integer id);
 }
