+++ transport/src/main/java/com/cooperative/transport/services/PanneService.java
@@ -22,7 +22,7 @@ public class PanneService {
     private final MotifPanneRepository motifPanneRepository;
     private final VoyageService voyageService;
 
-    public PanneDTO createPanne(Long voyageId, Long chauffeurId, String lieu, String motifPanneLibelle, String description, String photoUrl) {
+    public PanneDTO createPanne(Integer voyageId, Integer chauffeurId, String lieu, String motifPanneLibelle, String description, String photoUrl) {
         Optional<Voyages> voyageOpt = voyageRepository.findById(voyageId);
         if (voyageOpt.isEmpty() || !voyageOpt.get().getChauffeur().getId().equals(chauffeurId)) {
             throw new IllegalArgumentException("Voyage non trouvé ou accès refusé");
@@ -50,7 +50,7 @@ public class PanneService {
         return mapToPanneDTO(savedPanne);
     }
 
-    public List<PanneDTO> getPannesByChauffeur(Long chauffeurId) {
+    public List<PanneDTO> getPannesByChauffeur(Integer chauffeurId) {
         List<Pannes> pannes = panneRepository.findByChauffeurId(chauffeurId);
         return pannes.stream()
                 .map(this::mapToPanneDTO)
